refactor: Change project structure; update README
