--- src/filesystem.js
@@ -1,4 +1,0 @@
-const myFileSystem = {
-  file1: 'f1',
-  dir1: 'dir'
-};
