+++ src/index.html
@@ -42,9 +42,9 @@
         </p>
     </div>
     <!-- Global variables -->
-    <script src="filesystem.js"></script>
-    <script src="customCommands.js"></script>
+    <script src="js/filesystem.js"></script>
+    <script src="js/customCommands.js"></script>
     <!-- Minified build -->
-    <script src="bundle.min.js"></script>
+    <script src="js/bundle.min.js"></script>
 </body>
 </html>
