+++ src/main/app/Terminal.js
