+++ src/main/version.js
